+++ example/lib/basic_example.dart
@@ -3,7 +3,10 @@ import 'package:catcher/catcher_plugin.dart';
 
 main() {
   CatcherOptions debugOptions = CatcherOptions(PageReportMode(), [
-    EmailManualHandler(["recipient@email.com"]),
+    //EmailManualHandler(["recipient@email.com"]),
+    HttpHandler(HttpRequestType.post,
+        Uri.parse("https://jsonplaceholder.typicode.com/posts"),
+        printLogs: true),
     ConsoleHandler()
   ]);
   CatcherOptions releaseOptions = CatcherOptions(PageReportMode(), [